chore: move lang tests to a submodule
