+++ crates/lang/src/lib.rs
@@ -6,3 +6,6 @@ pub mod lexer;
 pub mod parser;
 pub mod tipo;
 pub mod token;
+
+#[cfg(test)]
+mod tests;
