+++ crates/lang/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod lexer;
+mod parser;
